+++ assets/js/admin.js
@@ -42,7 +42,7 @@ window.jQuery(document).ready(function ($) {
 
   // minimum requirements notice dismiss
   if (typeof mtmMinimumRequirementsNoticeAjax !== 'undefined' && mtmMinimumRequirementsNoticeAjax.ajax_url) {
-    $(document).on( 'click', '#matomo-minimumrequirements .notice-dismiss', function () {
+    $(document).on( 'click', '.matomo-minimum-requirements-notice .notice-dismiss', function () {
       $.post(mtmMinimumRequirementsNoticeAjax.ajax_url, {
         _ajax_nonce: mtmMinimumRequirementsNoticeAjax.nonce,
         action: 'matomo_minimum_requirements_notice_dismissed',