+++ matomo.php
@@ -5,6 +5,8 @@
  * Author: Matomo
  * Author URI: https://matomo.org
  * Version: 5.12.2
+ * Requires at least: 4.8
+ * Requires PHP: 7.2.5
  * Domain Path: /languages
  * WC requires at least: 2.4.0
  * WC tested up to: 11.0.1